Support account linking flows
Added logic so that the plugin can support account linking
flows. The selectEnrollee method will be called by COmanage
core. The method checks to see that the flow is being used for
account linking and if so allows the user to choose if she is
linking and creating an HC account. If so the user is prompted
for a password and then provisions into the HC IdP using
the existing username and the provided password.