Added support for identity linking flow with forced reauthentication
Added the function execute_plugin_selectEnrollee(). It detects when the
user is in the middle of an identity linking flow and adds a cookie
to signal that forced re-authentication is required. The cookie
signals to an enhanced app/webroot/auth/login/index.php that detects
it and then includes the right flag for the Shibboleth SP session
initiator to signal the IdP for forced re-authentication.